refactor: JpaRepository 들을 infra 패키지로 이동
