+++ src/main/java/kitchenpos/infra/JpaMenuGroupRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.infra;
 
+import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
