+++ src/main/java/kitchenpos/infra/JpaMenuRepository.java
@@ -1,8 +1,10 @@
-package kitchenpos.domain;
+package kitchenpos.infra;
 
 import jakarta.annotation.Nonnull;
 import java.util.List;
 import java.util.UUID;
+import kitchenpos.domain.Menu;
+import kitchenpos.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
