+++ src/main/java/kitchenpos/infra/JpaOrderRepository.java
@@ -1,6 +1,10 @@
-package kitchenpos.domain;
+package kitchenpos.infra;
 
 import jakarta.annotation.Nonnull;
+import kitchenpos.domain.Order;
+import kitchenpos.domain.OrderRepository;
+import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
