+++ src/main/java/kitchenpos/infra/JpaOrderTableRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.infra;
 
+import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
